+++ Sources/Main/Core/AnalyticsManager.swift
@@ -95,6 +95,7 @@ protocol AnalyticsManageable: AnyObject {
     /// <string>_rem_logout</string>
     /// <string>_rem_install</string>
     /// <string>_rem_visit</string>
+    /// <string>_rem_applink</string>
     /// <string>_rem_push_received</string>
     /// <string>_rem_push_notify</string>
     /// <string>_rem_push_auto_register</string>